+++ cpp/src/utilities/error_utils.hpp
@@ -63,6 +63,6 @@ struct logic_error : public std::logic_error {
   (!!(cond)) ? static_cast<void>(0)             \
              : throw cudf::detail::logic_error( \
                    "cuDF failure at: " __FILE__ \
-                   ": " CUDF_STRINGIFY(__LINE__) ". Reason: " reason)
+                   ":" CUDF_STRINGIFY(__LINE__) ". Reason: " reason)
 
 #endif